+++ R/tm_g_forest_rsp.R
@@ -185,6 +185,7 @@ template_forest_rsp <- function(dataname = "ANL",
 #' This teal module produces a grid style Forest plot for response data with ADaM structure.
 #'
 #' @inheritParams module_arguments
+#' @inheritParams tm_t_rsp
 #' @param fixed_symbol_size (`logical`)\cr When (`TRUE`), the same symbol size is used for plotting each
 #' estimate. Otherwise, the symbol size will be proportional to the sample size in each each subgroup.
 #'
@@ -198,7 +199,8 @@ template_forest_rsp <- function(dataname = "ANL",
 #' library(dplyr)
 #'
 #' ADSL <- synthetic_cdisc_data("latest")$adsl
-#' ADRS <- synthetic_cdisc_data("latest")$adrs
+#' ADRS <- synthetic_cdisc_data("latest")$adrs %>%
+#'   mutate(AVALC = d_onco_rsp_label(AVALC))
 #'
 #' arm_ref_comp = list(
 #'   ARM = list(
@@ -213,8 +215,12 @@ template_forest_rsp <- function(dataname = "ANL",
 #'
 #' app <- init(
 #'   data = cdisc_data(
-#'     cdisc_dataset("ADSL", ADSL, code = 'ADSL <- synthetic_cdisc_data("latest")$adsl'),
-#'     cdisc_dataset("ADRS", ADRS, code = 'ADRS <- synthetic_cdisc_data("latest")$adrs'),
+#'     cdisc_dataset("ADSL", ADSL),
+#'     cdisc_dataset("ADRS", ADRS),
+#'     code =
+#'       "ADSL <- synthetic_cdisc_data('latest')$adsl
+#'        ADRS <- synthetic_cdisc_data('latest')$adrs %>%
+#'          mutate(AVALC = d_onco_rsp_label(AVALC))",
 #'     check = TRUE
 #'   ),
 #'   modules = root_modules(
@@ -238,7 +244,20 @@ template_forest_rsp <- function(dataname = "ANL",
 #'         variable_choices(ADSL, c("STRATA1", "STRATA2")),
 #'         "STRATA2"
 #'       ),
-#'       plot_height = c(600L, 200L, 2000L)
+#'       plot_height = c(600L, 200L, 2000L),
+#'       default_responses = list(
+#'         BESRSPI = list(
+#'           rsp = c("Stable Disease (SD)", "Not Evaluable (NE)"),
+#'           levels = c("Complete Response (CR)", "Partial Response (PR)", "Stable Disease (SD)",
+#'                      "Progressive Disease (PD)", "Not Evaluable (NE)")),
+#'         INVET = list(
+#'           rsp = c("Complete Response (CR)", "Partial Response (PR)"),
+#'           levels = c("Complete Response (CR)", "Not Evaluable (NE)", "Partial Response (PR)",
+#'                      "Progressive Disease (PD)", "Stable Disease (SD)")),
+#'         OVRINV = list(
+#'           rsp = c("Progressive Disease (PD)", "Stable Disease (SD)"),
+#'           levels = c("Progressive Disease (PD)", "Stable Disease (SD)", "Not Evaluable (NE)"))
+#'       )
 #'     )
 #'   )
 #' )
@@ -258,6 +277,7 @@ tm_g_forest_rsp <- function(label,
                             strata_var,
                             fixed_symbol_size = TRUE,
                             conf_level = choices_selected(c(0.95, 0.9, 0.8), 0.95, keep_order = TRUE),
+                            default_responses = c("CR", "PR", "Y", "Complete Response (CR)", "Partial Response (PR)"),
                             plot_height = c(700L, 200L, 2000L),
                             plot_width = c(900L, 200L, 2000L),
                             pre_output = NULL,
@@ -282,6 +302,14 @@ tm_g_forest_rsp <- function(label,
   check_slider_input(plot_height, allow_null = FALSE)
   check_slider_input(plot_width)
 
+  assert_that(
+    is.list(default_responses) ||
+      is.null(default_responses) ||
+      is.character(default_responses) ||
+      is.numeric(default_responses),
+    msg = "`default_responses` must be a named list or an array."
+  )
+
   args <- as.list(environment())
 
   data_extract_list <- list(
@@ -304,6 +332,7 @@ tm_g_forest_rsp <- function(label,
         parentname = parentname,
         arm_ref_comp = arm_ref_comp,
         label = label,
+        default_responses = default_responses,
         plot_height = plot_height,
         plot_width = plot_width
       )
@@ -421,7 +450,8 @@ srv_g_forest_rsp <- function(input,
                              strata_var,
                              plot_height,
                              plot_width,
-                             label) {
+                             label,
+                             default_responses) {
   stopifnot(is_cdisc_data(datasets))
 
   init_chunks()
@@ -462,22 +492,31 @@ srv_g_forest_rsp <- function(input,
     anl_name = "ANL_ADSL"
   )
 
-  # Update UI choices depending on selection of previous options
-  observeEvent(anl_merged(), {
-    aval_var <- anl_merged()$columns_source$aval_var
-    if (nrow(anl_merged()$data()) == 0) {
-      responder_choices <- c("CR", "PR")
-      responder_sel <- c("CR", "PR")
-    } else {
-      responder_choices <- unique(anl_merged()$data()[[aval_var]])
-      responder_sel <- intersect(responder_choices, isolate(input$responders))
-    }
-    updateSelectInput(
-      session, "responders",
-      choices = responder_choices,
-      selected = responder_sel
-    )
-  })
+  observeEvent(
+    c(input[[extract_input("aval_var", "ADRS")]],
+      input[[extract_input("paramcd", paramcd$filter[[1]]$dataname, filter = TRUE)]]), {
+        aval_var <- anl_merged()$columns_source$aval_var
+        sel_param <- if (is.list(default_responses)) {
+          default_responses[[input[[extract_input("paramcd", paramcd$filter[[1]]$dataname, filter = TRUE)]]]]
+        } else default_responses
+        common_rsp <- if (is.list(sel_param)) {
+          sel_param$rsp
+        } else sel_param
+        responder_choices <- if (is_empty(aval_var)) {
+          character(0)
+        } else {
+          if ("levels" %in% names(sel_param)) {
+            if (length(intersect(unique(anl_merged()$data()[[aval_var]]), sel_param$levels)) > 1) {
+              sel_param$levels
+            } else union(unique(anl_merged()$data()[[aval_var]]), sel_param$levels)
+          } else unique(anl_merged()$data()[[aval_var]])
+        }
+        updateSelectInput(
+          session, "responders",
+          choices = responder_choices,
+          selected = intersect(responder_choices, common_rsp)
+        )
+    })
 
   # Prepare the analysis environment (filter data, check data, populate envir).
   validate_checks <- reactive({
@@ -518,10 +557,36 @@ srv_g_forest_rsp <- function(input,
       )
     }
 
+    if (!identical(default_responses, c("CR", "PR", "Y", "Complete Response (CR)", "Partial Response (PR)"))) {
+      validate(
+        need(all(unlist(lapply(default_responses, function(x) {
+          if (is.list(x) & "levels" %in% names(x)) {
+            lvls <- x$levels
+            all(x$rsp %in% lvls)
+          } else {
+            lvls <- unique(anl_merged()$data()[[input$`aval_var-dataset_ADRS_singleextract-select`]])
+            if ("rsp" %in% names(x)) {
+              all(x$rsp %in% lvls)
+            } else all(x %in% lvls)
+          }
+          }))),
+          "All selected default responses must be in the levels of AVAL.")
+      )
+    }
+
+    if (is.list(default_responses)) {
+      validate(
+        need(all(
+          grepl("\\.rsp|\\.levels", names(unlist(default_responses))) |
+            names(unlist(default_responses)) %in% names(default_responses)),
+          "The lists given for each AVAL in default_responses must be named 'rsp' and 'levels'.")
+      )
+    }
+
     validate(need(
       input$conf_level >= 0 && input$conf_level <= 1,
-      "Please choose a confidence level between 0 and 1"
-    ))
+      "Please choose a confidence level between 0 and 1")
+    )
 
     validate(
       need(is_character_single(input_aval_var), "Analysis variable should be a single column."),