+++ NEWS.md
@@ -11,9 +11,10 @@
 * Enhanced `tm_g_forest_rsp` to include selected parameter in title of the table. 
 * Enhanced `tm_g_forest_tte` to include selected parameter in title of the table. 
 
-
 ## Enhancements
 * Added support for logging with the `logger` package and added info level logs upon initialization of a module.
+* Added `default_responses` argument to `tm_t_binary_outcome`, `tm_t_rsp`, `tm_g_forest_rsp` to allow the user to specify default selected responses and possible response levels.
+* Updated `tm_t_binary_outcome` to show only the selected responses in the output table when "Show All Selected Response Categories" is checked.
 
 ### Miscellaneous
 * Updated R version requirement to >= 3.6.